+++ CreatesInputRockleGeometries.py
@@ -156,7 +156,8 @@ def createBlocks(cursor, inputBuildings, snappingTolerance = SNAPPING_TOLERANCE)
             AS {5}""".format(stackedBlockTable, ID_FIELD_STACKED_BLOCK, ID_FIELD_BLOCK,
                                 GEOM_FIELD, HEIGHT_FIELD, " UNION ALL ".join(listOfSqlQueries)))
     
-    # Drop intermediate tables
-    cursor.execute("DROP TABLE IF EXISTS {0}".format(",".join([correlTable])))
+    if not DEBUG:
+        # Drop intermediate tables
+        cursor.execute("DROP TABLE IF EXISTS {0}".format(",".join([correlTable])))
                         
     return blockTable, stackedBlockTable
